Changed the system for multi-dialog conversations
Now you create a chain of dialog gameobjects
that are either a list of stings, or a dialog option box
This requires the creation of more gameobjects
and specific scripts for each dialog, however:
There's less random method calling to try and get things to behave.
Hopefully this system is easier to work with.